+++ CHANGELOG.md
@@ -2,9 +2,6 @@
 
 ## Unreleased
 
-### Fixes
-
-- Add auto-start for session tracker when SDK is started after didBecomeActive (#5121)
 > [!Important]
 > This version creates new issue groups for your unhandled C++ exceptions because it now again reports the message of unhandled C++ exceptions, which we use for grouping.
 
@@ -20,6 +17,7 @@
 ### Fixes
 
 - Reporting unhandled C++ exception message (#5190)
+- Add auto-start for session tracker when SDK is started after didBecomeActive (#5121)
 
 ## 8.50.2